+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - importing from `bibtex` data defaults to **not** ignored non-standard entry types (#28)
 - suppress LaTex encoding warnings except when adding entries (#29)
+- `Search` and `Select` print warnings to the prompt while not implemented
 
 ### Fixed
 - `init` command ensures directory of database file exists